# Backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.BillingDashboardResponse:
    properties:
      billing:
        $ref: '#/definitions/dto.BillingDetails'
      invoices:
        items:
          $ref: '#/definitions/dto.InvoiceInfo'
        type: array
      subscription:
        $ref: '#/definitions/dto.SubscriptionBillingInfo'
      tenant:
        $ref: '#/definitions/dto.TenantBillingInfo'
      usage:
        $ref: '#/definitions/dto.UsageInfo'
    type: object
  dto.BillingDetails:
    properties:
      available_plans:
        items:
          $ref: '#/definitions/dto.PlanOption'
        type: array
      can_downgrade:
        type: boolean
      can_upgrade:
        type: boolean
      currency:
        type: string
      current_plan:
        type: string
      monthly_price:
        type: number
      next_billing:
        type: string
      payment_method:
        type: string
    type: object
  dto.CancelSubscriptionRequest:
    properties:
      reason:
        type: string
    type: object
...
  dto.InvoiceInfo:
    properties:
      amount:
        type: number
      download_url:
        type: string
      due_date:
        type: string
      id:
        type: string
      invoice_no:
        type: string
      issued_date:
        type: string
      paid_date:
        type: string
      status:
        type: string
    type: object
...
  dto.PlanOption:
    properties:
      description:
        type: string
      id:
        type: string
      is_current:
        type: boolean
      name:
        type: string
      price:
        type: number
      slug:
        type: string
    type: object
...
  dto.SubscriptionBillingInfo:
    properties:
      auto_renew:
        type: boolean
      days_left:
        type: integer
      end_date:
        type: string
      id:
        type: string
      is_trial:
        type: boolean
      next_billing_date:
        type: string
      payment_method:
        type: string
      plan_id:
        type: string
      plan_name:
        type: string
      plan_slug:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
  dto.TenantBillingInfo:
    properties:
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  dto.UpdatePaymentMethodRequest:
    properties:
      card_holder:
        type: string
      card_number:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      payment_method:
        type: string
    required:
    - payment_method
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      auto_renew:
        type: boolean
      payment_method:
        type: string
      plan_id:
        type: string
    required:
    - plan_id
    type: object
...
  dto.UpdateTenantSettingsRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dto.UsageInfo:
    properties:
      current_period_end:
        type: string
      current_period_start:
        type: string
      days_remaining:
        type: integer
      days_used:
        type: integer
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: localhost:8089
info:
  contact:
    email: support@rtrwnet.com
    name: RT RW Net Support
  description: |-
    API untuk sistem manajemen RT RW Net berbasis SaaS dengan multi-tenant architecture.

    ## Features
    - Multi-tenant architecture
    - JWT authentication
    - Free trial support (7 days)
    - Subscription management
    - Billing dashboard

    ## Authentication
    Gunakan Bearer token di header Authorization untuk protected endpoints.
    Untuk tenant identification, sertakan X-Tenant-ID di header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: RT RW Net SaaS Backend API
  version: 2.0.0
paths:
...
  /billing:
    get:
      consumes:
      - application/json
      description: Retrieve complete billing information including tenant details,
        current subscription, usage statistics, available plans, and recent invoices.
        Requires authentication and tenant ID.
      produces:
      - application/json
      responses:
        "200":
          description: Billing dashboard retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BillingDashboardResponse'
              type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No subscription found for tenant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - TenantID: []
      summary: Get billing dashboard
      tags:
      - Billing
  /billing/cancel:
    post:
      consumes:
      - application/json
      description: Cancel current subscription. Subscription will be marked as cancelled
        and auto-renewal will be disabled. Optionally provide cancellation reason.
      parameters:
      - description: Cancellation request with optional reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CancelSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription cancelled successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No subscription found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - TenantID: []
      summary: Cancel subscription
      tags:
      - Billing
  /billing/payment-method:
    put:
      consumes:
      - application/json
      description: Update payment method for subscription. Can include credit card
        details or other payment method information.
      parameters:
      - description: Payment method data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePaymentMethodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment method updated successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Validation error - payment method required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No subscription found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - TenantID: []
      summary: Update payment method
      tags:
      - Billing
  /billing/settings:
    put:
      consumes:
      - application/json
      description: Update tenant information such as name, email, or phone number.
      parameters:
      - description: Tenant settings data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTenantSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Settings updated successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Tenant not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - TenantID: []
      summary: Update tenant settings
      tags:
      - Billing
  /billing/subscription:
    put:
      consumes:
      - application/json
      description: Change subscription plan, payment method, or auto-renewal settings.
        If upgrading from trial, subscription will be activated immediately.
      parameters:
      - description: Subscription update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription updated successfully
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid plan or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: No subscription found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - TenantID: []
      summary: Update subscription plan
      tags:
      - Billing
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  TenantID:
    description: Tenant ID for multi-tenant access
    in: header
    name: X-Tenant-ID
    type: apiKey
swagger: "2.0"